+++ tiberiandawn/ccini.cpp
@@ -736,10 +736,10 @@ bool CCINIClass::Put_TerrainType(char const* section, char const* entry, Terrain
  * HISTORY:                                                                                    *
  *   07/11/1996 JLB : Created.                                                                 *
  *=============================================================================================*/
-long CCINIClass::Get_Buildings(char const* section, char const* entry, long defvalue) const
+int CCINIClass::Get_Buildings(char const* section, char const* entry, int defvalue) const
 {
     char buffer[128];
-    long pre;
+    int pre;
 
     if (Get_String(section, entry, "", buffer, sizeof(buffer))) {
 
@@ -781,7 +781,7 @@ long CCINIClass::Get_Buildings(char const* section, char const* entry, long defv
  * HISTORY:                                                                                    *
  *   07/11/1996 JLB : Created.                                                                 *
  *=============================================================================================*/
-bool CCINIClass::Put_Buildings(char const* section, char const* entry, long value)
+bool CCINIClass::Put_Buildings(char const* section, char const* entry, int value)
 {
     char buffer[128] = "";
     int maxi = (32 < STRUCT_COUNT) ? 32 : STRUCT_COUNT;
